Moving email subscription to homepage
