+++ src/components/index.ts
@@ -1,0 +1,4 @@
+export { default as EmailInput } from "./emailInput";
+export { default as Layout } from "./layout";
+export { default as SearchBar } from "./searchBar";
+export { default as SEO } from "./seo";
